Complete splitting off object index
